- extract outcomeBalance as fn tradesOutcomeBalance on trades - extract collateralAmountUSDSpent as fn tradesCollateralAmountUSDSpent on trades
